Updating client to support ID and name sent for new users
When a new user is added to the group a new. Using the find method we look to see if the user already exists in the list or not. If not we push the new user in and send out an update to the group via the websocket as normal.

Also making use of the new msgData property to send specific data along to the other connected clients.

The combination of find and filter makes it really easy to look up objects in a list by ID and then removing, so the removeUser function remain relatively simple.